Major rewrite of indexed part selects.
This patch is a major rewrite of the indexed part selects (+: and -:).

It made the following enhancements:

1. Make indexed part selects work correctly with both big and little
   endian vectors.

2. Add a warning flag that warns about constant out of bounds/or 'bx
   indexed selects.

3. Moved the -: parameter code to its own routine.

4. Added support for straddling before part selects in a CA.

5. Added more assert(! number_is_unknown) statements.

6. Add warning for &PV<> select with a signed index signal that is
  less than the width of an int. This will be fixed later.

7. Add support for loading a 'bx/'bz constant into a numeric register.

8. Add a number of signed value fixes to the compiler/code generator.

9. Major fix of draw_select_expr() in the code generator.
(cherry picked from commit 2b17366ad5d59b458042c15311c3cb5bcd467553)

+++ netlist.h
@@ -3829,8 +3829,8 @@ class NetESignal  : public NetExpr {
       const NetNet* sig() const;
       NetNet* sig();
 	// Declared vector dimensions for the signal.
-      unsigned msi() const;
-      unsigned lsi() const;
+      long msi() const;
+      long lsi() const;
 
       virtual ivl_variable_type_t expr_type() const;
